reorganized and added two more endpoints
