+++ Nessie_Python_SDK/urlConstants.py
